Update Release Script and Doc for News Changes

+++ docs/news.py
@@ -23,8 +23,9 @@ def run(*argv):
 
 
 def release(args):
+    print(args)
     version_info = VersionInfo()
-    if not version_info.is_release:
+    if not args.solo_test and not version_info.is_release:
         sys.exit('ERROR: Not a release')
     version_entry = '## Version '
     news_md_title = version_entry + version_info.ver + ' of OpenDDS'
@@ -33,24 +34,33 @@ def release(args):
     news_rst_path = releases_path / (version_info.v_ver + '.rst')
     release_notes_path = docs_path / 'gh-release-notes.md'
 
+    needs_changes = False
+    print('Checking files...')
+    if not news_rst_path.is_file():
+        print('  Missing ' + news_rst_path.name)
+        needs_changes = True
+    if not release_notes_path.is_file():
+        print('  Missing ' + release_notes_path.name)
+        needs_changes = True
+    with news_md_path.open('r') as f:
+        for line in f.readlines():
+            line = line.rstrip()
+            if line.startswith(version_entry):
+                if line != news_md_title:
+                    print('  Missing entry in NEWS.md for ' + version_info.ver)
+                    needs_changes = True
+                break
     news = parse_newsd()
     if news.empty():
-        print('No news fragments, checking for files...')
-        if not news_rst_path.is_file():
-            sys.exit('ERROR: news.d is empty and missing ' + news_rst_path.name + '!')
-        if not release_notes_path.is_file():
-            sys.exit('ERROR: news.d is empty and missing ' + release_notes_path.name + '!')
+        if not args.mock and needs_changes:
+            sys.exit('ERROR: news.d is empty!')
+    elif not args.mock:
+        needs_changes = True
 
-        print('No news fragments, checking for NEWS.md entry...')
-        with news_md_path.open('r') as f:
-            for line in f.readlines():
-                line = line.rstrip()
-                if line.startswith(version_entry):
-                    if line == news_md_title:
-                        break
-                    else:
-                        sys.exit('ERROR: news.d is empty and missing NEWS.md entry!')
-    else:
+    if needs_changes:
+        if not args.remedy:
+            sys.exit('ERROR: --remedy needs to be passed to process the news')
+
         news.print_all()
         with news_rst_path.open('w') as f:
             news.print_all(file=f)
@@ -65,7 +75,7 @@ def release(args):
         print(this_release_md)
         # That was for NEWS.md, now generate GitHub release notes
         release_notes, dl_count = re.subn(r'^Download \[this release.*$', '', this_release_md, flags=re.MULTILINE)
-        if dl_count != 1:
+        if not args.solo_test and dl_count != 1:
             sys.exit('ERROR: Expected 1 download lines in release notes, got ' + str(dl_count))
         release_notes = re.sub(r'\n\n+', r'\n\n', release_notes)
         with (docs_path / 'gh-release-notes.md').open('w') as f:
@@ -92,11 +102,14 @@ def release(args):
                     print(this_release_md, file=f)
                     inserted = True
                 print(line, file=f)
+    else:
+        print('  No changes needed!')
 
-        print('Removing fragments...')
-        for fragment in get_fragments():
-            print(fragment)
-            fragment.unlink()
+        if not args.solo_test:
+            print('Removing fragments...')
+            for fragment in get_fragments():
+                print(fragment)
+                fragment.unlink()
 
     return 0
 
@@ -116,6 +129,12 @@ if __name__ == '__main__':
     subparsers = arg_parser.add_subparsers(required=True)
 
     arg_parser_release = subparsers.add_parser('release')
+    arg_parser_release.add_argument('--remedy', action='store_true',
+        help='Allow changes to the news')
+    arg_parser_release.add_argument('--mock', action='store_true',
+        help='Allow empty news.d')
+    arg_parser_release.add_argument('--solo-test', action='store_true',
+        help='Allow a non-release test that does not remove the fragments')
     arg_parser_release.set_defaults(func=release)
 
     arg_parser_preview = subparsers.add_parser('preview')
